moving toNTSC() functions to command.go
