+++ master/internal/command/command.go
@@ -26,7 +26,11 @@ import (
 	"github.com/determined-ai/determined/master/pkg/protoutils"
 	"github.com/determined-ai/determined/master/pkg/tasks"
 	"github.com/determined-ai/determined/proto/pkg/apiv1"
+	"github.com/determined-ai/determined/proto/pkg/commandv1"
+	"github.com/determined-ai/determined/proto/pkg/notebookv1"
+	"github.com/determined-ai/determined/proto/pkg/shellv1"
 	"github.com/determined-ai/determined/proto/pkg/taskv1"
+	"github.com/determined-ai/determined/proto/pkg/tensorboardv1"
 )
 
 // terminatedDuration defines the amount of time the command stays in a
@@ -303,6 +307,103 @@ func (c *command) deleteIfInWorkspace(req *apiv1.DeleteWorkspaceRequest) {
 		if err != nil {
 			c.syslog.WithError(err).Warn("failed to kill allocation while deleting workspace")
 		}
+	}
+}
+
+// toCommand() takes a *command from the command service registry & returns a *commandv1.Command.
+func (c *command) toCommand() *commandv1.Command {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	allo := c.refreshAllocationState()
+	return &commandv1.Command{
+		Id:           c.stringID(),
+		State:        enrichState(allo.State),
+		Description:  c.Config.Description,
+		Container:    allo.SingleContainer().ToProto(),
+		StartTime:    protoutils.ToTimestamp(c.registeredTime),
+		Username:     c.Base.Owner.Username,
+		UserId:       int32(c.Base.Owner.ID),
+		DisplayName:  c.Base.Owner.DisplayName.ValueOrZero(),
+		ResourcePool: c.Config.Resources.ResourcePool,
+		ExitStatus:   c.exitStatus.String(),
+		JobId:        c.jobID.String(),
+		WorkspaceId:  int32(c.GenericCommandSpec.Metadata.WorkspaceID),
+	}
+}
+
+// toNotebook() takes a *command from the command service registry & returns a *notebookv1.Notebook.
+func (c *command) toNotebook() *notebookv1.Notebook {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	allo := c.refreshAllocationState()
+	return &notebookv1.Notebook{
+		Id:             c.stringID(),
+		State:          enrichState(allo.State),
+		Description:    c.Config.Description,
+		Container:      allo.SingleContainer().ToProto(),
+		ServiceAddress: c.serviceAddress(),
+		StartTime:      protoutils.ToTimestamp(c.registeredTime),
+		Username:       c.Base.Owner.Username,
+		UserId:         int32(c.Base.Owner.ID),
+		DisplayName:    c.Base.Owner.DisplayName.ValueOrZero(),
+		ResourcePool:   c.Config.Resources.ResourcePool,
+		ExitStatus:     c.exitStatus.String(),
+		JobId:          c.jobID.String(),
+		WorkspaceId:    int32(c.GenericCommandSpec.Metadata.WorkspaceID),
+	}
+}
+
+// toShell() takes a *command from the command service registry & returns a *shellv1.Shell.
+func (c *command) toShell() *shellv1.Shell {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	allo := c.refreshAllocationState()
+	return &shellv1.Shell{
+		Id:             c.stringID(),
+		State:          enrichState(allo.State),
+		Description:    c.Config.Description,
+		StartTime:      protoutils.ToTimestamp(c.registeredTime),
+		Container:      allo.SingleContainer().ToProto(),
+		PrivateKey:     *c.Metadata.PrivateKey,
+		PublicKey:      *c.Metadata.PublicKey,
+		Username:       c.Base.Owner.Username,
+		UserId:         int32(c.Base.Owner.ID),
+		DisplayName:    c.Base.Owner.DisplayName.ValueOrZero(),
+		ResourcePool:   c.Config.Resources.ResourcePool,
+		ExitStatus:     c.exitStatus.String(),
+		Addresses:      toProto(allo.SingleContainerAddresses()),
+		AgentUserGroup: protoutils.ToStruct(c.Base.AgentUserGroup),
+		JobId:          c.jobID.String(),
+		WorkspaceId:    int32(c.GenericCommandSpec.Metadata.WorkspaceID),
+	}
+}
+
+// toTensorboard() takes a *command from the command service registry & returns a *tensorboardv1.Tensorboard.
+func (c *command) toTensorboard() *tensorboardv1.Tensorboard {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	allo := c.refreshAllocationState()
+	state := enrichState(allo.State)
+	return &tensorboardv1.Tensorboard{
+		Id:             c.stringID(),
+		State:          state,
+		Description:    c.Config.Description,
+		StartTime:      protoutils.ToTimestamp(c.registeredTime),
+		Container:      allo.SingleContainer().ToProto(),
+		ServiceAddress: c.serviceAddress(),
+		ExperimentIds:  c.Metadata.ExperimentIDs,
+		TrialIds:       c.Metadata.TrialIDs,
+		Username:       c.Base.Owner.Username,
+		UserId:         int32(c.Base.Owner.ID),
+		DisplayName:    c.Base.Owner.DisplayName.ValueOrZero(),
+		ResourcePool:   c.Config.Resources.ResourcePool,
+		ExitStatus:     c.exitStatus.String(),
+		JobId:          c.jobID.String(),
+		WorkspaceId:    int32(c.GenericCommandSpec.Metadata.WorkspaceID),
 	}
 }
